requirements.txt
Flask==2.1.2
gunicorn==20.1.0
flake8>=4.0.0
flask_script==2.0.5
mongoengine==0.24.1
flask_testing==0.8.1
python-dotenv==0.20.0
bcrypt==3.2.2
grpcio-tools==1.47.0
futures==3.1.1
grpcio-reflection==1.47.0
